pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["shu <shu@ceshmina.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.3"
google-cloud-firestore = "^2.10.0"
flask-login = "^0.6.2"
gunicorn = "^20.1.0"
openai = "^0.27.0"
flask-cors = "^3.0.10"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
flake8 = "^6.0.0"
mypy = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
